Settings: Add the PersistantSettings class to hold all Settings
We use a Singleton here. On the first call of the Instance() method the
Settings get either loaded from Disk or default-initialized. Subsequent
calls just return the instance. Upon quitting, the GameManager calls the
SaveToDisk method to save the contents to a persistant path. On my
windows machine this ends up in
%AppData%\LocalLow\DefaultCompany\UGS_game\usersettings.json.
Changing the Company and gamename should be done at some point.